Update to EN translation
Didn't notice your message, here are the artifact part labels.

I will be working on the "Ornament" translations, the ornaments seems to be furnishing items. This will take me awhile to do as I wont be able to spawn them in-game without going in the Teapot, since the Teapot locations currently have no solid ground to see what the in-game translations are, also too lazy to see if there is a solid ground somewhere.